Mejoras en el flujo de la aplicacion y cambios en el navbar y sidebar. Modularizacion del codigo en createUser
