+++ FrontEnd/src/Components/Navbar/Navbar.css
@@ -6,20 +6,23 @@
   width: 12vh;
   height: 12vh;
   margin-top: -6px;
+  margin-left: 1vh;
 }
 
 .activeProject{
   margin-top: auto;
   margin-bottom: auto;
-  padding-left: 1vh;
-  padding-right: 1vh;
+  margin-left: 1vh;
+  /* padding-left: 1vh;
+  padding-right: 1vh; */
   text-align: center;
   line-height: 7vh;
   color: white;
-  border-radius: 15px;
+  border-radius: 8px;
   font-size: 2vh;
+  /* border: 0.2px solid white; */
   /* text-decoration: underline; */
-  background: #083148;
+  /* background: #082f48; */
   height: 7vh;
 }
 
@@ -27,19 +30,27 @@
   margin-top: auto;
   margin-bottom: auto;
   text-align: center;
-  margin-right: -10vh;
-  line-height: 7vh;
+  /* margin-right: -10vh; */
+  line-height: 6.5vh;
   color: white;
   border-radius: 15px;
   font-size: 3vh;
   border: 1px solid white;
-  background: #083148;
+  background: #133c54;
   border-radius: 50px;
-  height: 7vh;
-  width: 7vh;
+  height: 6.5vh;
+  width: 6.5vh;
 }
 
 .navbar-project-user{
   display: flex;
   gap: 25px;
+  background: #083148;
+  padding-left: 1vh;
+  padding-right: 1vh;
+  border-radius: 50px;
+  height: 7.8vh;
+  margin-top: auto;
+  margin-bottom: auto;
+  margin-right: -10vh;
 }
