+++ FrontEnd/src/Pages/payrollProjects/SelectProject.js
@@ -44,7 +44,7 @@ const SelectProject = () => {
   const handleProjectSelection = (projectName) => {
     console.log(`El seleccionado:${projectName}`);
     dispatch(updateActiveProject(projectName));
-    navigate('/');
+    navigate('/employees');
 
   }
 
@@ -71,7 +71,6 @@ const SelectProject = () => {
               <div key={project.Nombre} className='project-projectBox'>
 
                 <button
-
                   onClick={() => handleProjectSelection(project.Nombre)} className='project-projectLogo'>
                   {project.Nombre.charAt(0).toLocaleUpperCase()}
                 </button>
