+++ FrontEnd/src/Slices/user/requests/postLogin.js
@@ -1,24 +1,5 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 
-// export const postLogin = createAsyncThunk('usuarios/postLogin', async (credentials) => {
-//     const loginFetch = await fetch('https://api.ticolitas.com/usuarios/login', 
-//     {
-//         method: 'POST',
-//         headers: {
-//             "Content-type": "application/json",
-//         },
-//         body: JSON.stringify({
-//             email: credentials.username,
-//             password: credentials.password,
-//         }),
-//     });
-
-//     const userData = await loginFetch.json();
-
-//     return getData(userData, loginFetch);
-
-// });
-
 export const postLogin = createAsyncThunk('usuarios/postLogin', async (credentials) => {
     const loginFetch = await fetch('http://localhost:4000/users', 
     {
@@ -27,7 +8,7 @@ export const postLogin = createAsyncThunk('usuarios/postLogin', async (credentia
             "Content-type": "application/json",
         },
         body: JSON.stringify({
-            Email: credentials.username,
+            Email: credentials.email,
             Contrasenia: credentials.password,
         }),
     });
